Move package `io.element.android.features.messages.impl.mentions` to `io.element.android.features.messages.impl.messagecomposer.suggestions`
